+++ 4Week_Mission/src/main/java/com/ll/exam/Week_Mission/app/security/config/ApiSecurityConfig.java
@@ -1,17 +1,15 @@
 package com.ll.exam.Week_Mission.app.security.config;
 
-import com.ll.exam.Week_Mission.app.jwt.filter.JwtAuthorizationFilter;
+import com.ll.exam.Week_Mission.app.security.jwt.filter.JwtAuthorizationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.cors.reactive.CorsUtils;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
